requirements.txt
python-dotenv
boto3
openai
docx